fix(test): Integer not BigInteger for SQLite autoincrement id
Test fixtures define scope_change_events with raw Column DDL (separate
from the ORM model). SQLite only treats a column as the ROWID alias
(autoincrementing) when the type is exactly INTEGER PRIMARY KEY; with
BIGINT PRIMARY KEY the column is regular and never auto-fills, which
tripped NOT NULL on every INSERT that omitted id.

Switching to Integer in the test fixtures lets SQLite autoincrement.
The production schema is BIGSERIAL via Alembic (Postgres) and is
unchanged.

This unmasks three separate cross-dialect bugs in test_snapshots
that were hidden by the earlier IntegrityError crash:
  - test_case_b_first_sighting_pre_start_is_silent
  - test_case_c_mid_sprint_addition_records_event
  - test_existing_sp_change_emits_event
Investigated separately.

+++ backend/tests/test_projects_freeze.py
@@ -155,7 +155,9 @@ def _create_subset(sync_conn) -> None:
     Table(
         "scope_change_events",
         md,
-        Column("id", BigInteger, primary_key=True, autoincrement=True),
+        # Integer (not BigInteger) so SQLite treats it as ROWID alias and
+        # autoincrements; production schema is BIGSERIAL via Alembic.
+        Column("id", Integer, primary_key=True, autoincrement=True),
         Column("issue_key", String, nullable=False),
         Column("sprint_name", String, nullable=False),
         Column("change_type", String, nullable=False),
